Keep guards sorted
This is a fairly large commit. The std::vectors we were using to store guards
have been changed to std::sets. These std::sets take a GuardSetComparator that
keeps the set sorted. The GuardSetComparator works by using the
InternalKeyComparator in the VersionStorageInfo.

The main change we will have to make going forward is remembering to not use
the [] notation for getting guards from the new_guards_ and complete_guards_
std::unordered_maps. Using [] in std::unordered_map requires there to be a
default constructor for the std::set, but the std::set is parameterized as
std::set<GuardMetaData, GuardSetComparator>>, and GuardSetComparator
intentionally does not have a default constructor since it needs access to the
VersionStorageInfo class to get the internal key comparator. If you get a weird
template error when compiling, it is most likely because of this reason.