Move GuardSet to sep. header file; keep output guards in Compaction
This should fix a race condition caused by compaction happening at the same time
as Put(). The race condition was that the set of new guards could change during
compaction, and we were checking the set of new guards once during key-value
compaction and once during installing the compaction results into a new Version,
which meant that the set of output guards could change in the meantime.

This change makes it so the Compaction class computes the set of output guards
when it is constructed. This is making the assumption that Compaction is
constructed while the DB mutex is held, so it should be thread-safe.